+++ tests/EllipseTest.php
@@ -4,14 +4,17 @@ declare(strict_types=1);
 
 namespace UIAwesome\Html\Svg\Tests;
 
+use InvalidArgumentException;
 use PHPUnit\Framework\Attributes\Group;
 use PHPUnit\Framework\TestCase;
 use UIAwesome\Html\Core\Factory\SimpleFactory;
 use UIAwesome\Html\Core\Values\{Aria, DataProperty, Language, Role};
+use UIAwesome\Html\Helper\Enum;
+use UIAwesome\Html\Helper\Exception\Message;
 use UIAwesome\Html\Svg\Ellipse;
 use UIAwesome\Html\Svg\Tests\Support\Stub\DefaultProvider;
 use UIAwesome\Html\Svg\Tests\Support\TestSupport;
-use UIAwesome\Html\Svg\Values\{FillRule, StrokeLineCap, StrokeLineJoin};
+use UIAwesome\Html\Svg\Values\{FillRule, StrokeLineCap, StrokeLineJoin, SvgProperty};
 
 /**
  * Test suite for {@see Ellipse} element functionality and behavior.
@@ -565,5 +568,97 @@ final class EllipseTest extends TestCase
             $ellipse->transform(''),
             'Should return a new instance when setting the attribute, ensuring immutability.',
         );
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidFillOpacityValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_OUT_OF_RANGE_OR_NULL->getMessage(0, 1),
+        );
+
+        Ellipse::tag()->fillOpacity('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidFillRuleValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            Message::VALUE_NOT_IN_LIST->getMessage(
+                'invalid-value',
+                SvgProperty::FILL_RULE->value,
+                implode('\', \'', Enum::normalizeArray(FillRule::cases())),
+            ),
+        );
+
+        Ellipse::tag()->fillRule('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidOpacityValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_OUT_OF_RANGE_OR_NULL->getMessage(0, 1),
+        );
+
+        Ellipse::tag()->opacity('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidPathLengthValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_MUST_BE_POSITIVE_NUMBER_OR_NULL->getMessage(),
+        );
+
+        Ellipse::tag()->pathLength('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidStrokeLineCapValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            Message::VALUE_NOT_IN_LIST->getMessage(
+                'invalid-value',
+                SvgProperty::STROKE_LINECAP->value,
+                implode('\', \'', Enum::normalizeArray(StrokeLineCap::cases())),
+            ),
+        );
+
+        Ellipse::tag()->strokeLineCap('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidStrokeLineJoinValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            Message::VALUE_NOT_IN_LIST->getMessage(
+                'invalid-value',
+                SvgProperty::STROKE_LINEJOIN->value,
+                implode('\', \'', Enum::normalizeArray(StrokeLineJoin::cases())),
+            ),
+        );
+
+        Ellipse::tag()->strokeLineJoin('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidStrokeMiterlimitValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_MUST_BE_GTE_ONE_OR_NULL->getMessage(),
+        );
+
+        Ellipse::tag()->strokeMiterlimit('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidStrokeOpacityValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_OUT_OF_RANGE_OR_NULL->getMessage('0', '1'),
+        );
+
+        Ellipse::tag()->strokeOpacity('invalid-value');
     }
 }