+++ tests/SvgTest.php
@@ -10,11 +10,12 @@ use PHPUnit\Framework\TestCase;
 use RuntimeException;
 use UIAwesome\Html\Core\Factory\SimpleFactory;
 use UIAwesome\Html\Core\Values\{Aria, ContentEditable, DataProperty, Direction, Draggable, Language, Role, Translate};
-use UIAwesome\Html\Svg\Exception\Message;
+use UIAwesome\Html\Helper\Enum;
+use UIAwesome\Html\Helper\Exception\Message;
 use UIAwesome\Html\Svg\Svg;
 use UIAwesome\Html\Svg\Tests\Support\Stub\{DefaultProvider, DefaultThemeProvider};
 use UIAwesome\Html\Svg\Tests\Support\TestSupport;
-use UIAwesome\Html\Svg\Values\{FillRule, StrokeLineCap, StrokeLineJoin};
+use UIAwesome\Html\Svg\Values\{FillRule, StrokeLineCap, StrokeLineJoin, SvgProperty};
 
 /**
  * Test suite for {@see Svg} element functionality and behavior.
@@ -1042,7 +1043,7 @@ final class SvgTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage(
-            Message::CONTENT_AND_FILEPATH_CANNOT_BE_BOTH_EMPTY->getMessage(),
+            \UIAwesome\Html\Svg\Exception\Message::CONTENT_AND_FILEPATH_CANNOT_BE_BOTH_EMPTY->getMessage(),
         );
 
         Svg::tag()->content('')->render();
@@ -1052,17 +1053,99 @@ final class SvgTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage(
-            Message::CONTENT_AND_FILEPATH_CANNOT_BE_BOTH_EMPTY->getMessage(),
+            \UIAwesome\Html\Svg\Exception\Message::CONTENT_AND_FILEPATH_CANNOT_BE_BOTH_EMPTY->getMessage(),
         );
 
         Svg::tag()->filePath('')->render();
     }
 
-    public function testThrowInvalidArgumentExceptionForInvalidValueTitleAttribute(): void
+    public function testThrowInvalidArgumentExceptionForSettingInvalidFillOpacityValue(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage(
-            Message::TITLE_ATTRIBUTE_MUST_BE_STRING_OR_NULL->getMessage(),
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_OUT_OF_RANGE_OR_NULL->getMessage(0, 1),
+        );
+
+        Svg::tag()->fillOpacity('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidFillRuleValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            Message::VALUE_NOT_IN_LIST->getMessage(
+                'invalid-value',
+                SvgProperty::FILL_RULE->value,
+                implode('\', \'', Enum::normalizeArray(FillRule::cases())),
+            ),
+        );
+
+        Svg::tag()->fillRule('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidOpacityValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_OUT_OF_RANGE_OR_NULL->getMessage(0, 1),
+        );
+
+        Svg::tag()->opacity('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidStrokeLineCapValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            Message::VALUE_NOT_IN_LIST->getMessage(
+                'invalid-value',
+                SvgProperty::STROKE_LINECAP->value,
+                implode('\', \'', Enum::normalizeArray(StrokeLineCap::cases())),
+            ),
+        );
+
+        Svg::tag()->strokeLineCap('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidStrokeLineJoinValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            Message::VALUE_NOT_IN_LIST->getMessage(
+                'invalid-value',
+                SvgProperty::STROKE_LINEJOIN->value,
+                implode('\', \'', Enum::normalizeArray(StrokeLineJoin::cases())),
+            ),
+        );
+
+        Svg::tag()->strokeLineJoin('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidStrokeMiterlimitValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_MUST_BE_GTE_ONE_OR_NULL->getMessage(),
+        );
+
+        Svg::tag()->strokeMiterlimit('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidStrokeOpacityValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_OUT_OF_RANGE_OR_NULL->getMessage('0', '1'),
+        );
+
+        Svg::tag()->strokeOpacity('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingInvalidTitleValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::TITLE_ATTRIBUTE_MUST_BE_STRING_OR_NULL->getMessage(),
         );
 
         Svg::tag()->attributes(['title' => []])->content('value')->render();
@@ -1074,7 +1157,7 @@ final class SvgTest extends TestCase
 
         $this->expectException(RuntimeException::class);
         $this->expectExceptionMessage(
-            Message::FAILED_TO_READ_FILE->getMessage($file),
+            \UIAwesome\Html\Svg\Exception\Message::FAILED_TO_READ_FILE->getMessage($file),
         );
 
         Svg::tag()->filePath($file)->render();
@@ -1086,7 +1169,7 @@ final class SvgTest extends TestCase
 
         $this->expectException(RuntimeException::class);
         $this->expectExceptionMessage(
-            Message::FAILED_TO_SANITIZE_SVG->getMessage($filePath),
+            \UIAwesome\Html\Svg\Exception\Message::FAILED_TO_SANITIZE_SVG->getMessage($filePath),
         );
 
         Svg::tag()->filePath($filePath)->render();